refact(models): creacion de un archivo unico por cada modelo
