+++ src/models/activity.js
@@ -1,11 +1,5 @@
 const { mongoose } = require('../config/db')
 
-const userSchema = new mongoose.Schema({
-  name: String,
-  email: { type: String, unique: true },
-  password: String
-  // registerDate: Date.now()
-})
 const activitySchema = new mongoose.Schema({
   title: String,
   description: String,
@@ -18,9 +12,5 @@ const activitySchema = new mongoose.Schema({
 })
 
 const Activity = mongoose.model('Activity', activitySchema)
-const User = mongoose.model('User', userSchema)
 
-module.exports = {
-  Activity,
-  User
-}
+module.exports = Activity
