+++ src/services/users.js
@@ -1,4 +1,4 @@
-const { User } = require('../models/index')
+const { User } = require('../models/user')
 
 class Users {
   async getOneByEmail (email) {
